- optimized ballSim

+++ Constants.h
@@ -1,6 +1,6 @@
 #pragma once
 
-const int screenWidth = 1600;
-const int screenHeight = 900;
+const double screenWidth = 1600;
+const double screenHeight = 900;
 const int bigG = 5000;
 const float M_PIE = std::acos(-1.0);
